Arrange your data.frames

Re-arrange your data.frame in ascending or descending order given one or several columns.

Usage

# arrange_(df, ...)
# desange_(df, ...)

Arguments

df

A data.frame.

...

A formula used for arranging the data.frame (e.g., ~c(col1, col2)).

Value

The functions return an object of the same type as df. Properties:

Columns are not modified.

Output rows are in the order specified by the formula.

Data frame attributes are preserved.

Examples

tmp <- arrange_(CO2, ~c(conc))
head(tmp)
#>   Plant   Type  Treatment conc uptake
#> 1   Qn1 Quebec nonchilled   95   16.0
#> 2   Qn2 Quebec nonchilled   95   13.6
#> 3   Qn3 Quebec nonchilled   95   16.2
#> 4   Qc1 Quebec    chilled   95   14.2
#> 5   Qc2 Quebec    chilled   95    9.3
#> 6   Qc3 Quebec    chilled   95   15.1

co2 <- m61r(df = CO2)
co2$arrange(~c(conc))
co2$head()
#>   Plant   Type  Treatment conc uptake
#> 1   Qn1 Quebec nonchilled   95   16.0
#> 2   Qn2 Quebec nonchilled   95   13.6
#> 3   Qn3 Quebec nonchilled   95   16.2
#> 4   Qc1 Quebec    chilled   95   14.2
#> 5   Qc2 Quebec    chilled   95    9.3
#> 6   Qc3 Quebec    chilled   95   15.1