Evaluate Formula Expressions on Data Subsets

The core engine for Non-Standard Evaluation within m61r. expression_ evaluates a user-provided formula within the context of a data frame, optionally for calculated groups. The result relies on the Base R functions with and eval.

Usage

# expression_(df, group_info = NULL, fun_expr)

Arguments

df

data.frame

group_info

An optional list of grouping indices and keys, typically generated by get_group_indices_(). If NULL, the expression runs over the entire df.

fun_expr

A formula (~<expression>) that describes the R code to be executed.

Value

The function returns a list.

If group_info is NULL, the list contains the result of fun_expr executed on the entire df.

If group_info is provided, the list contains the results of fun_expr executed on each group's subset of the df.

Examples

# Non-Grouped Evaluation (for mutate)
expression_(CO2,fun_expr=~conc/uptake)
#> $fun_expr
#>  [1]  5.937500  5.756579  7.183908  9.408602 14.164306 17.219388 25.188917
#>  [8]  6.985294  6.410256  6.738544  8.373206 12.315271 16.304348 22.573363
#> [15]  5.864198  5.401235  6.203474  8.313539 11.655012 15.375854 21.978022
#> [22]  6.690141  7.261411  8.250825 10.115607 15.384615 19.067797 25.839793
#> [29] 10.215054  6.410256  7.142857  9.020619 12.953368 18.000000 23.584906
#> [36]  6.291391  8.333333  6.561680 10.294118 12.853470 17.045455 24.154589
#> [43]  8.962264  9.114583  9.541985 11.666667 16.181230 20.833333 28.169014
#> [50]  7.916667  7.954545  8.169935 11.006289 15.432099 21.704180 31.746032
#> [57]  8.407080  9.020619  9.689922 12.544803 17.543860 24.021352 35.971223
#> [64]  9.047619 11.744966 13.812155 18.518519 25.641026 30.405405 45.662100
#> [71] 12.337662 15.350877 20.325203 26.923077 40.000000 49.270073 69.444444
#> [78]  8.962264  9.722222 13.966480 19.553073 27.932961 35.714286 50.251256
#> 

# Grouped Evaluation (for summarise)
group_info <- get_group_indices_(CO2, ~Type)
expression_(CO2, group_info = group_info, fun_expr=~mean(uptake))
#> $fun_expr
#> $fun_expr$Quebec
#> [1] 33.54286
#> 
#> $fun_expr$Mississippi
#> [1] 20.88333
#> 
#> 

# Complex Grouped Evaluation (results in a list per group)
expression_(CO2, group_info = group_info, fun_expr=~lm(uptake~conc))
#> $fun_expr
#> $fun_expr$Quebec
#> 
#> Call:
#> lm(formula = uptake ~ conc)
#> 
#> Coefficients:
#> (Intercept)         conc  
#>    23.50304      0.02308  
#> 
#> 
#> $fun_expr$Mississippi
#> 
#> Call:
#> lm(formula = uptake ~ conc)
#> 
#> Coefficients:
#> (Intercept)         conc  
#>    15.49754      0.01238  
#> 
#> 
#>