Create m61r object

Create a m61r object that enables to run a sequence of operations on a data.frame.

Usage

# m61r(df = NULL)

# S3 method for class 'm61r'
x[i, j, ...]

# S3 method for class 'm61r'
x[i, j] <- value

# S3 method for class 'm61r'
print(x, ...)

# S3 method for class 'm61r'
names(x, ...)

# S3 method for class 'm61r'
dim(x, ...)

# S3 method for class 'm61r'
as.data.frame(x, ...)

# S3 method for class 'm61r'
rbind(x, ...)

# S3 method for class 'm61r'
cbind(x, ...)

Arguments

df

data.frame

x

object of class m61r

i

row

j

column

...

further arguments passed to or from other methods

value

value to be assigned

Value

The function m61r returns an object of type m61r.

Argument df get stored internally to the object m61r. One manipulates the internal data.frame by using internal functions similar to the ones implemented in package m61r for data.frames as arrange, desange, filter, join and its relatives, mutate and transmutate, gather and spread, select, group_by, summarise, values and modify. The result of the last action is stored internally to the object m61r until the internal function values get called. It is thus possible to create a readable sequence of actions on a data.frame.

In addition,

[.m61r returns a subset of the internal data.frame embedded to the object m61r.

[<-.m61r assigns value to the internal data.frame embedded to the object m61r.

print.m61r prints the internal data.frame embedded to the object m61r.

names.m61r provides the names of the column of the internal data.frame embedded to the object m61r.

dim.m61r provides the dimensions of the internal data.frame embedded to the object m61r.

as.data.frame.m61r extracts the internal data.frame embedded to the object m61r.

cbind.m61r combines by columns two objects m61r.

rbind.m61r combines by rows two objects m61r.

left_join, anti_join, full_join, inner_join, right_join, semi_join join two objects m61r.

The m61r object maintains an internal state. For advanced transformations, users can use .SD() within formulas. .SD() (Subset of Data) refers to the current data.frame being processed. If a group_by operation is active, .SD() refers to the current group only.

Finally, it is possible to clone a m61r object into a new one by using the internal function clone.

Examples

  # init
  co2 <- m61r(df=CO2)

  # filter
  co2$filter(~Plant=="Qn1")
  co2
#>   Plant   Type  Treatment conc uptake
#> 1   Qn1 Quebec nonchilled   95   16.0
#> 2   Qn1 Quebec nonchilled  175   30.4
#> 3   Qn1 Quebec nonchilled  250   34.8
#> 4   Qn1 Quebec nonchilled  350   37.2
#> 5   Qn1 Quebec nonchilled  500   35.3
#> 6   Qn1 Quebec nonchilled  675   39.2
#> 7   Qn1 Quebec nonchilled 1000   39.7

  co2$filter(~Type=="Quebec")
  co2
#>    Plant   Type  Treatment conc uptake
#> 1    Qn1 Quebec nonchilled   95   16.0
#> 2    Qn1 Quebec nonchilled  175   30.4
#> 3    Qn1 Quebec nonchilled  250   34.8
#> 4    Qn1 Quebec nonchilled  350   37.2
#> 5    Qn1 Quebec nonchilled  500   35.3
#> 6    Qn1 Quebec nonchilled  675   39.2
#> 7    Qn1 Quebec nonchilled 1000   39.7
#> 8    Qn2 Quebec nonchilled   95   13.6
#> 9    Qn2 Quebec nonchilled  175   27.3
#> 10   Qn2 Quebec nonchilled  250   37.1
#> 11   Qn2 Quebec nonchilled  350   41.8
#> 12   Qn2 Quebec nonchilled  500   40.6
#> 13   Qn2 Quebec nonchilled  675   41.4
#> 14   Qn2 Quebec nonchilled 1000   44.3
#> 15   Qn3 Quebec nonchilled   95   16.2
#> 16   Qn3 Quebec nonchilled  175   32.4
#> 17   Qn3 Quebec nonchilled  250   40.3
#> 18   Qn3 Quebec nonchilled  350   42.1
#> 19   Qn3 Quebec nonchilled  500   42.9
#> 20   Qn3 Quebec nonchilled  675   43.9
#> 21   Qn3 Quebec nonchilled 1000   45.5
#> 22   Qc1 Quebec    chilled   95   14.2
#> 23   Qc1 Quebec    chilled  175   24.1
#> 24   Qc1 Quebec    chilled  250   30.3
#> 25   Qc1 Quebec    chilled  350   34.6
#> 26   Qc1 Quebec    chilled  500   32.5
#> 27   Qc1 Quebec    chilled  675   35.4
#> 28   Qc1 Quebec    chilled 1000   38.7
#> 29   Qc2 Quebec    chilled   95    9.3
#> 30   Qc2 Quebec    chilled  175   27.3
#> 31   Qc2 Quebec    chilled  250   35.0
#> 32   Qc2 Quebec    chilled  350   38.8
#> 33   Qc2 Quebec    chilled  500   38.6
#> 34   Qc2 Quebec    chilled  675   37.5
#> 35   Qc2 Quebec    chilled 1000   42.4
#> 36   Qc3 Quebec    chilled   95   15.1
#> 37   Qc3 Quebec    chilled  175   21.0
#> 38   Qc3 Quebec    chilled  250   38.1
#> 39   Qc3 Quebec    chilled  350   34.0
#> 40   Qc3 Quebec    chilled  500   38.9
#> 41   Qc3 Quebec    chilled  675   39.6
#> 42   Qc3 Quebec    chilled 1000   41.4

  # select
  co2$select(~Type)
  co2
#>           Type
#> 1       Quebec
#> 2       Quebec
#> 3       Quebec
#> 4       Quebec
#> 5       Quebec
#> 6       Quebec
#> 7       Quebec
#> 8       Quebec
#> 9       Quebec
#> 10      Quebec
#> 11      Quebec
#> 12      Quebec
#> 13      Quebec
#> 14      Quebec
#> 15      Quebec
#> 16      Quebec
#> 17      Quebec
#> 18      Quebec
#> 19      Quebec
#> 20      Quebec
#> 21      Quebec
#> 22      Quebec
#> 23      Quebec
#> 24      Quebec
#> 25      Quebec
#> 26      Quebec
#> 27      Quebec
#> 28      Quebec
#> 29      Quebec
#> 30      Quebec
#> 31      Quebec
#> 32      Quebec
#> 33      Quebec
#> 34      Quebec
#> 35      Quebec
#> 36      Quebec
#> 37      Quebec
#> 38      Quebec
#> 39      Quebec
#> 40      Quebec
#> 41      Quebec
#> 42      Quebec
#> 43 Mississippi
#> 44 Mississippi
#> 45 Mississippi
#> 46 Mississippi
#> 47 Mississippi
#> 48 Mississippi
#> 49 Mississippi
#> 50 Mississippi
#> 51 Mississippi
#> 52 Mississippi
#> 53 Mississippi
#> 54 Mississippi
#> 55 Mississippi
#> 56 Mississippi
#> 57 Mississippi
#> 58 Mississippi
#> 59 Mississippi
#> 60 Mississippi
#> 61 Mississippi
#> 62 Mississippi
#> 63 Mississippi
#> 64 Mississippi
#> 65 Mississippi
#> 66 Mississippi
#> 67 Mississippi
#> 68 Mississippi
#> 69 Mississippi
#> 70 Mississippi
#> 71 Mississippi
#> 72 Mississippi
#> 73 Mississippi
#> 74 Mississippi
#> 75 Mississippi
#> 76 Mississippi
#> 77 Mississippi
#> 78 Mississippi
#> 79 Mississippi
#> 80 Mississippi
#> 81 Mississippi
#> 82 Mississippi
#> 83 Mississippi
#> 84 Mississippi

  co2$select(~c(Plant,Type))
  co2
#>    Plant        Type
#> 1    Qn1      Quebec
#> 2    Qn1      Quebec
#> 3    Qn1      Quebec
#> 4    Qn1      Quebec
#> 5    Qn1      Quebec
#> 6    Qn1      Quebec
#> 7    Qn1      Quebec
#> 8    Qn2      Quebec
#> 9    Qn2      Quebec
#> 10   Qn2      Quebec
#> 11   Qn2      Quebec
#> 12   Qn2      Quebec
#> 13   Qn2      Quebec
#> 14   Qn2      Quebec
#> 15   Qn3      Quebec
#> 16   Qn3      Quebec
#> 17   Qn3      Quebec
#> 18   Qn3      Quebec
#> 19   Qn3      Quebec
#> 20   Qn3      Quebec
#> 21   Qn3      Quebec
#> 22   Qc1      Quebec
#> 23   Qc1      Quebec
#> 24   Qc1      Quebec
#> 25   Qc1      Quebec
#> 26   Qc1      Quebec
#> 27   Qc1      Quebec
#> 28   Qc1      Quebec
#> 29   Qc2      Quebec
#> 30   Qc2      Quebec
#> 31   Qc2      Quebec
#> 32   Qc2      Quebec
#> 33   Qc2      Quebec
#> 34   Qc2      Quebec
#> 35   Qc2      Quebec
#> 36   Qc3      Quebec
#> 37   Qc3      Quebec
#> 38   Qc3      Quebec
#> 39   Qc3      Quebec
#> 40   Qc3      Quebec
#> 41   Qc3      Quebec
#> 42   Qc3      Quebec
#> 43   Mn1 Mississippi
#> 44   Mn1 Mississippi
#> 45   Mn1 Mississippi
#> 46   Mn1 Mississippi
#> 47   Mn1 Mississippi
#> 48   Mn1 Mississippi
#> 49   Mn1 Mississippi
#> 50   Mn2 Mississippi
#> 51   Mn2 Mississippi
#> 52   Mn2 Mississippi
#> 53   Mn2 Mississippi
#> 54   Mn2 Mississippi
#> 55   Mn2 Mississippi
#> 56   Mn2 Mississippi
#> 57   Mn3 Mississippi
#> 58   Mn3 Mississippi
#> 59   Mn3 Mississippi
#> 60   Mn3 Mississippi
#> 61   Mn3 Mississippi
#> 62   Mn3 Mississippi
#> 63   Mn3 Mississippi
#> 64   Mc1 Mississippi
#> 65   Mc1 Mississippi
#> 66   Mc1 Mississippi
#> 67   Mc1 Mississippi
#> 68   Mc1 Mississippi
#> 69   Mc1 Mississippi
#> 70   Mc1 Mississippi
#> 71   Mc2 Mississippi
#> 72   Mc2 Mississippi
#> 73   Mc2 Mississippi
#> 74   Mc2 Mississippi
#> 75   Mc2 Mississippi
#> 76   Mc2 Mississippi
#> 77   Mc2 Mississippi
#> 78   Mc3 Mississippi
#> 79   Mc3 Mississippi
#> 80   Mc3 Mississippi
#> 81   Mc3 Mississippi
#> 82   Mc3 Mississippi
#> 83   Mc3 Mississippi
#> 84   Mc3 Mississippi

  co2$select(~-Type)
  co2
#>    Plant  Treatment conc uptake
#> 1    Qn1 nonchilled   95   16.0
#> 2    Qn1 nonchilled  175   30.4
#> 3    Qn1 nonchilled  250   34.8
#> 4    Qn1 nonchilled  350   37.2
#> 5    Qn1 nonchilled  500   35.3
#> 6    Qn1 nonchilled  675   39.2
#> 7    Qn1 nonchilled 1000   39.7
#> 8    Qn2 nonchilled   95   13.6
#> 9    Qn2 nonchilled  175   27.3
#> 10   Qn2 nonchilled  250   37.1
#> 11   Qn2 nonchilled  350   41.8
#> 12   Qn2 nonchilled  500   40.6
#> 13   Qn2 nonchilled  675   41.4
#> 14   Qn2 nonchilled 1000   44.3
#> 15   Qn3 nonchilled   95   16.2
#> 16   Qn3 nonchilled  175   32.4
#> 17   Qn3 nonchilled  250   40.3
#> 18   Qn3 nonchilled  350   42.1
#> 19   Qn3 nonchilled  500   42.9
#> 20   Qn3 nonchilled  675   43.9
#> 21   Qn3 nonchilled 1000   45.5
#> 22   Qc1    chilled   95   14.2
#> 23   Qc1    chilled  175   24.1
#> 24   Qc1    chilled  250   30.3
#> 25   Qc1    chilled  350   34.6
#> 26   Qc1    chilled  500   32.5
#> 27   Qc1    chilled  675   35.4
#> 28   Qc1    chilled 1000   38.7
#> 29   Qc2    chilled   95    9.3
#> 30   Qc2    chilled  175   27.3
#> 31   Qc2    chilled  250   35.0
#> 32   Qc2    chilled  350   38.8
#> 33   Qc2    chilled  500   38.6
#> 34   Qc2    chilled  675   37.5
#> 35   Qc2    chilled 1000   42.4
#> 36   Qc3    chilled   95   15.1
#> 37   Qc3    chilled  175   21.0
#> 38   Qc3    chilled  250   38.1
#> 39   Qc3    chilled  350   34.0
#> 40   Qc3    chilled  500   38.9
#> 41   Qc3    chilled  675   39.6
#> 42   Qc3    chilled 1000   41.4
#> 43   Mn1 nonchilled   95   10.6
#> 44   Mn1 nonchilled  175   19.2
#> 45   Mn1 nonchilled  250   26.2
#> 46   Mn1 nonchilled  350   30.0
#> 47   Mn1 nonchilled  500   30.9
#> 48   Mn1 nonchilled  675   32.4
#> 49   Mn1 nonchilled 1000   35.5
#> 50   Mn2 nonchilled   95   12.0
#> 51   Mn2 nonchilled  175   22.0
#> 52   Mn2 nonchilled  250   30.6
#> 53   Mn2 nonchilled  350   31.8
#> 54   Mn2 nonchilled  500   32.4
#> 55   Mn2 nonchilled  675   31.1
#> 56   Mn2 nonchilled 1000   31.5
#> 57   Mn3 nonchilled   95   11.3
#> 58   Mn3 nonchilled  175   19.4
#> 59   Mn3 nonchilled  250   25.8
#> 60   Mn3 nonchilled  350   27.9
#> 61   Mn3 nonchilled  500   28.5
#> 62   Mn3 nonchilled  675   28.1
#> 63   Mn3 nonchilled 1000   27.8
#> 64   Mc1    chilled   95   10.5
#> 65   Mc1    chilled  175   14.9
#> 66   Mc1    chilled  250   18.1
#> 67   Mc1    chilled  350   18.9
#> 68   Mc1    chilled  500   19.5
#> 69   Mc1    chilled  675   22.2
#> 70   Mc1    chilled 1000   21.9
#> 71   Mc2    chilled   95    7.7
#> 72   Mc2    chilled  175   11.4
#> 73   Mc2    chilled  250   12.3
#> 74   Mc2    chilled  350   13.0
#> 75   Mc2    chilled  500   12.5
#> 76   Mc2    chilled  675   13.7
#> 77   Mc2    chilled 1000   14.4
#> 78   Mc3    chilled   95   10.6
#> 79   Mc3    chilled  175   18.0
#> 80   Mc3    chilled  250   17.9
#> 81   Mc3    chilled  350   17.9
#> 82   Mc3    chilled  500   17.9
#> 83   Mc3    chilled  675   18.9
#> 84   Mc3    chilled 1000   19.9

  co2$select(variable=~-(Plant:Treatment))
  co2
#>    conc uptake
#> 1    95   16.0
#> 2   175   30.4
#> 3   250   34.8
#> 4   350   37.2
#> 5   500   35.3
#> 6   675   39.2
#> 7  1000   39.7
#> 8    95   13.6
#> 9   175   27.3
#> 10  250   37.1
#> 11  350   41.8
#> 12  500   40.6
#> 13  675   41.4
#> 14 1000   44.3
#> 15   95   16.2
#> 16  175   32.4
#> 17  250   40.3
#> 18  350   42.1
#> 19  500   42.9
#> 20  675   43.9
#> 21 1000   45.5
#> 22   95   14.2
#> 23  175   24.1
#> 24  250   30.3
#> 25  350   34.6
#> 26  500   32.5
#> 27  675   35.4
#> 28 1000   38.7
#> 29   95    9.3
#> 30  175   27.3
#> 31  250   35.0
#> 32  350   38.8
#> 33  500   38.6
#> 34  675   37.5
#> 35 1000   42.4
#> 36   95   15.1
#> 37  175   21.0
#> 38  250   38.1
#> 39  350   34.0
#> 40  500   38.9
#> 41  675   39.6
#> 42 1000   41.4
#> 43   95   10.6
#> 44  175   19.2
#> 45  250   26.2
#> 46  350   30.0
#> 47  500   30.9
#> 48  675   32.4
#> 49 1000   35.5
#> 50   95   12.0
#> 51  175   22.0
#> 52  250   30.6
#> 53  350   31.8
#> 54  500   32.4
#> 55  675   31.1
#> 56 1000   31.5
#> 57   95   11.3
#> 58  175   19.4
#> 59  250   25.8
#> 60  350   27.9
#> 61  500   28.5
#> 62  675   28.1
#> 63 1000   27.8
#> 64   95   10.5
#> 65  175   14.9
#> 66  250   18.1
#> 67  350   18.9
#> 68  500   19.5
#> 69  675   22.2
#> 70 1000   21.9
#> 71   95    7.7
#> 72  175   11.4
#> 73  250   12.3
#> 74  350   13.0
#> 75  500   12.5
#> 76  675   13.7
#> 77 1000   14.4
#> 78   95   10.6
#> 79  175   18.0
#> 80  250   17.9
#> 81  350   17.9
#> 82  500   17.9
#> 83  675   18.9
#> 84 1000   19.9

  # mutate/transmutate
  co2$mutate(z=~conc/uptake)
  co2
#>    Plant        Type  Treatment conc uptake         z
#> 1    Qn1      Quebec nonchilled   95   16.0  5.937500
#> 2    Qn1      Quebec nonchilled  175   30.4  5.756579
#> 3    Qn1      Quebec nonchilled  250   34.8  7.183908
#> 4    Qn1      Quebec nonchilled  350   37.2  9.408602
#> 5    Qn1      Quebec nonchilled  500   35.3 14.164306
#> 6    Qn1      Quebec nonchilled  675   39.2 17.219388
#> 7    Qn1      Quebec nonchilled 1000   39.7 25.188917
#> 8    Qn2      Quebec nonchilled   95   13.6  6.985294
#> 9    Qn2      Quebec nonchilled  175   27.3  6.410256
#> 10   Qn2      Quebec nonchilled  250   37.1  6.738544
#> 11   Qn2      Quebec nonchilled  350   41.8  8.373206
#> 12   Qn2      Quebec nonchilled  500   40.6 12.315271
#> 13   Qn2      Quebec nonchilled  675   41.4 16.304348
#> 14   Qn2      Quebec nonchilled 1000   44.3 22.573363
#> 15   Qn3      Quebec nonchilled   95   16.2  5.864198
#> 16   Qn3      Quebec nonchilled  175   32.4  5.401235
#> 17   Qn3      Quebec nonchilled  250   40.3  6.203474
#> 18   Qn3      Quebec nonchilled  350   42.1  8.313539
#> 19   Qn3      Quebec nonchilled  500   42.9 11.655012
#> 20   Qn3      Quebec nonchilled  675   43.9 15.375854
#> 21   Qn3      Quebec nonchilled 1000   45.5 21.978022
#> 22   Qc1      Quebec    chilled   95   14.2  6.690141
#> 23   Qc1      Quebec    chilled  175   24.1  7.261411
#> 24   Qc1      Quebec    chilled  250   30.3  8.250825
#> 25   Qc1      Quebec    chilled  350   34.6 10.115607
#> 26   Qc1      Quebec    chilled  500   32.5 15.384615
#> 27   Qc1      Quebec    chilled  675   35.4 19.067797
#> 28   Qc1      Quebec    chilled 1000   38.7 25.839793
#> 29   Qc2      Quebec    chilled   95    9.3 10.215054
#> 30   Qc2      Quebec    chilled  175   27.3  6.410256
#> 31   Qc2      Quebec    chilled  250   35.0  7.142857
#> 32   Qc2      Quebec    chilled  350   38.8  9.020619
#> 33   Qc2      Quebec    chilled  500   38.6 12.953368
#> 34   Qc2      Quebec    chilled  675   37.5 18.000000
#> 35   Qc2      Quebec    chilled 1000   42.4 23.584906
#> 36   Qc3      Quebec    chilled   95   15.1  6.291391
#> 37   Qc3      Quebec    chilled  175   21.0  8.333333
#> 38   Qc3      Quebec    chilled  250   38.1  6.561680
#> 39   Qc3      Quebec    chilled  350   34.0 10.294118
#> 40   Qc3      Quebec    chilled  500   38.9 12.853470
#> 41   Qc3      Quebec    chilled  675   39.6 17.045455
#> 42   Qc3      Quebec    chilled 1000   41.4 24.154589
#> 43   Mn1 Mississippi nonchilled   95   10.6  8.962264
#> 44   Mn1 Mississippi nonchilled  175   19.2  9.114583
#> 45   Mn1 Mississippi nonchilled  250   26.2  9.541985
#> 46   Mn1 Mississippi nonchilled  350   30.0 11.666667
#> 47   Mn1 Mississippi nonchilled  500   30.9 16.181230
#> 48   Mn1 Mississippi nonchilled  675   32.4 20.833333
#> 49   Mn1 Mississippi nonchilled 1000   35.5 28.169014
#> 50   Mn2 Mississippi nonchilled   95   12.0  7.916667
#> 51   Mn2 Mississippi nonchilled  175   22.0  7.954545
#> 52   Mn2 Mississippi nonchilled  250   30.6  8.169935
#> 53   Mn2 Mississippi nonchilled  350   31.8 11.006289
#> 54   Mn2 Mississippi nonchilled  500   32.4 15.432099
#> 55   Mn2 Mississippi nonchilled  675   31.1 21.704180
#> 56   Mn2 Mississippi nonchilled 1000   31.5 31.746032
#> 57   Mn3 Mississippi nonchilled   95   11.3  8.407080
#> 58   Mn3 Mississippi nonchilled  175   19.4  9.020619
#> 59   Mn3 Mississippi nonchilled  250   25.8  9.689922
#> 60   Mn3 Mississippi nonchilled  350   27.9 12.544803
#> 61   Mn3 Mississippi nonchilled  500   28.5 17.543860
#> 62   Mn3 Mississippi nonchilled  675   28.1 24.021352
#> 63   Mn3 Mississippi nonchilled 1000   27.8 35.971223
#> 64   Mc1 Mississippi    chilled   95   10.5  9.047619
#> 65   Mc1 Mississippi    chilled  175   14.9 11.744966
#> 66   Mc1 Mississippi    chilled  250   18.1 13.812155
#> 67   Mc1 Mississippi    chilled  350   18.9 18.518519
#> 68   Mc1 Mississippi    chilled  500   19.5 25.641026
#> 69   Mc1 Mississippi    chilled  675   22.2 30.405405
#> 70   Mc1 Mississippi    chilled 1000   21.9 45.662100
#> 71   Mc2 Mississippi    chilled   95    7.7 12.337662
#> 72   Mc2 Mississippi    chilled  175   11.4 15.350877
#> 73   Mc2 Mississippi    chilled  250   12.3 20.325203
#> 74   Mc2 Mississippi    chilled  350   13.0 26.923077
#> 75   Mc2 Mississippi    chilled  500   12.5 40.000000
#> 76   Mc2 Mississippi    chilled  675   13.7 49.270073
#> 77   Mc2 Mississippi    chilled 1000   14.4 69.444444
#> 78   Mc3 Mississippi    chilled   95   10.6  8.962264
#> 79   Mc3 Mississippi    chilled  175   18.0  9.722222
#> 80   Mc3 Mississippi    chilled  250   17.9 13.966480
#> 81   Mc3 Mississippi    chilled  350   17.9 19.553073
#> 82   Mc3 Mississippi    chilled  500   17.9 27.932961
#> 83   Mc3 Mississippi    chilled  675   18.9 35.714286
#> 84   Mc3 Mississippi    chilled 1000   19.9 50.251256

  #co2$mutate(mean=~mean(uptake))
  #co2
  #Warning message:
  #In FUN(X[[i]], ...) : Expression mean has incompatible length.

  co2$mutate(z1=~uptake/conc,y=~conc/100)
  co2
#>    Plant        Type  Treatment conc uptake         z1     y
#> 1    Qn1      Quebec nonchilled   95   16.0 0.16842105  0.95
#> 2    Qn1      Quebec nonchilled  175   30.4 0.17371429  1.75
#> 3    Qn1      Quebec nonchilled  250   34.8 0.13920000  2.50
#> 4    Qn1      Quebec nonchilled  350   37.2 0.10628571  3.50
#> 5    Qn1      Quebec nonchilled  500   35.3 0.07060000  5.00
#> 6    Qn1      Quebec nonchilled  675   39.2 0.05807407  6.75
#> 7    Qn1      Quebec nonchilled 1000   39.7 0.03970000 10.00
#> 8    Qn2      Quebec nonchilled   95   13.6 0.14315789  0.95
#> 9    Qn2      Quebec nonchilled  175   27.3 0.15600000  1.75
#> 10   Qn2      Quebec nonchilled  250   37.1 0.14840000  2.50
#> 11   Qn2      Quebec nonchilled  350   41.8 0.11942857  3.50
#> 12   Qn2      Quebec nonchilled  500   40.6 0.08120000  5.00
#> 13   Qn2      Quebec nonchilled  675   41.4 0.06133333  6.75
#> 14   Qn2      Quebec nonchilled 1000   44.3 0.04430000 10.00
#> 15   Qn3      Quebec nonchilled   95   16.2 0.17052632  0.95
#> 16   Qn3      Quebec nonchilled  175   32.4 0.18514286  1.75
#> 17   Qn3      Quebec nonchilled  250   40.3 0.16120000  2.50
#> 18   Qn3      Quebec nonchilled  350   42.1 0.12028571  3.50
#> 19   Qn3      Quebec nonchilled  500   42.9 0.08580000  5.00
#> 20   Qn3      Quebec nonchilled  675   43.9 0.06503704  6.75
#> 21   Qn3      Quebec nonchilled 1000   45.5 0.04550000 10.00
#> 22   Qc1      Quebec    chilled   95   14.2 0.14947368  0.95
#> 23   Qc1      Quebec    chilled  175   24.1 0.13771429  1.75
#> 24   Qc1      Quebec    chilled  250   30.3 0.12120000  2.50
#> 25   Qc1      Quebec    chilled  350   34.6 0.09885714  3.50
#> 26   Qc1      Quebec    chilled  500   32.5 0.06500000  5.00
#> 27   Qc1      Quebec    chilled  675   35.4 0.05244444  6.75
#> 28   Qc1      Quebec    chilled 1000   38.7 0.03870000 10.00
#> 29   Qc2      Quebec    chilled   95    9.3 0.09789474  0.95
#> 30   Qc2      Quebec    chilled  175   27.3 0.15600000  1.75
#> 31   Qc2      Quebec    chilled  250   35.0 0.14000000  2.50
#> 32   Qc2      Quebec    chilled  350   38.8 0.11085714  3.50
#> 33   Qc2      Quebec    chilled  500   38.6 0.07720000  5.00
#> 34   Qc2      Quebec    chilled  675   37.5 0.05555556  6.75
#> 35   Qc2      Quebec    chilled 1000   42.4 0.04240000 10.00
#> 36   Qc3      Quebec    chilled   95   15.1 0.15894737  0.95
#> 37   Qc3      Quebec    chilled  175   21.0 0.12000000  1.75
#> 38   Qc3      Quebec    chilled  250   38.1 0.15240000  2.50
#> 39   Qc3      Quebec    chilled  350   34.0 0.09714286  3.50
#> 40   Qc3      Quebec    chilled  500   38.9 0.07780000  5.00
#> 41   Qc3      Quebec    chilled  675   39.6 0.05866667  6.75
#> 42   Qc3      Quebec    chilled 1000   41.4 0.04140000 10.00
#> 43   Mn1 Mississippi nonchilled   95   10.6 0.11157895  0.95
#> 44   Mn1 Mississippi nonchilled  175   19.2 0.10971429  1.75
#> 45   Mn1 Mississippi nonchilled  250   26.2 0.10480000  2.50
#> 46   Mn1 Mississippi nonchilled  350   30.0 0.08571429  3.50
#> 47   Mn1 Mississippi nonchilled  500   30.9 0.06180000  5.00
#> 48   Mn1 Mississippi nonchilled  675   32.4 0.04800000  6.75
#> 49   Mn1 Mississippi nonchilled 1000   35.5 0.03550000 10.00
#> 50   Mn2 Mississippi nonchilled   95   12.0 0.12631579  0.95
#> 51   Mn2 Mississippi nonchilled  175   22.0 0.12571429  1.75
#> 52   Mn2 Mississippi nonchilled  250   30.6 0.12240000  2.50
#> 53   Mn2 Mississippi nonchilled  350   31.8 0.09085714  3.50
#> 54   Mn2 Mississippi nonchilled  500   32.4 0.06480000  5.00
#> 55   Mn2 Mississippi nonchilled  675   31.1 0.04607407  6.75
#> 56   Mn2 Mississippi nonchilled 1000   31.5 0.03150000 10.00
#> 57   Mn3 Mississippi nonchilled   95   11.3 0.11894737  0.95
#> 58   Mn3 Mississippi nonchilled  175   19.4 0.11085714  1.75
#> 59   Mn3 Mississippi nonchilled  250   25.8 0.10320000  2.50
#> 60   Mn3 Mississippi nonchilled  350   27.9 0.07971429  3.50
#> 61   Mn3 Mississippi nonchilled  500   28.5 0.05700000  5.00
#> 62   Mn3 Mississippi nonchilled  675   28.1 0.04162963  6.75
#> 63   Mn3 Mississippi nonchilled 1000   27.8 0.02780000 10.00
#> 64   Mc1 Mississippi    chilled   95   10.5 0.11052632  0.95
#> 65   Mc1 Mississippi    chilled  175   14.9 0.08514286  1.75
#> 66   Mc1 Mississippi    chilled  250   18.1 0.07240000  2.50
#> 67   Mc1 Mississippi    chilled  350   18.9 0.05400000  3.50
#> 68   Mc1 Mississippi    chilled  500   19.5 0.03900000  5.00
#> 69   Mc1 Mississippi    chilled  675   22.2 0.03288889  6.75
#> 70   Mc1 Mississippi    chilled 1000   21.9 0.02190000 10.00
#> 71   Mc2 Mississippi    chilled   95    7.7 0.08105263  0.95
#> 72   Mc2 Mississippi    chilled  175   11.4 0.06514286  1.75
#> 73   Mc2 Mississippi    chilled  250   12.3 0.04920000  2.50
#> 74   Mc2 Mississippi    chilled  350   13.0 0.03714286  3.50
#> 75   Mc2 Mississippi    chilled  500   12.5 0.02500000  5.00
#> 76   Mc2 Mississippi    chilled  675   13.7 0.02029630  6.75
#> 77   Mc2 Mississippi    chilled 1000   14.4 0.01440000 10.00
#> 78   Mc3 Mississippi    chilled   95   10.6 0.11157895  0.95
#> 79   Mc3 Mississippi    chilled  175   18.0 0.10285714  1.75
#> 80   Mc3 Mississippi    chilled  250   17.9 0.07160000  2.50
#> 81   Mc3 Mississippi    chilled  350   17.9 0.05114286  3.50
#> 82   Mc3 Mississippi    chilled  500   17.9 0.03580000  5.00
#> 83   Mc3 Mississippi    chilled  675   18.9 0.02800000  6.75
#> 84   Mc3 Mississippi    chilled 1000   19.9 0.01990000 10.00

  co2$transmutate(z2=~uptake/conc,y2=~conc/100)
  co2
#>            z2    y2
#> 1  0.16842105  0.95
#> 2  0.17371429  1.75
#> 3  0.13920000  2.50
#> 4  0.10628571  3.50
#> 5  0.07060000  5.00
#> 6  0.05807407  6.75
#> 7  0.03970000 10.00
#> 8  0.14315789  0.95
#> 9  0.15600000  1.75
#> 10 0.14840000  2.50
#> 11 0.11942857  3.50
#> 12 0.08120000  5.00
#> 13 0.06133333  6.75
#> 14 0.04430000 10.00
#> 15 0.17052632  0.95
#> 16 0.18514286  1.75
#> 17 0.16120000  2.50
#> 18 0.12028571  3.50
#> 19 0.08580000  5.00
#> 20 0.06503704  6.75
#> 21 0.04550000 10.00
#> 22 0.14947368  0.95
#> 23 0.13771429  1.75
#> 24 0.12120000  2.50
#> 25 0.09885714  3.50
#> 26 0.06500000  5.00
#> 27 0.05244444  6.75
#> 28 0.03870000 10.00
#> 29 0.09789474  0.95
#> 30 0.15600000  1.75
#> 31 0.14000000  2.50
#> 32 0.11085714  3.50
#> 33 0.07720000  5.00
#> 34 0.05555556  6.75
#> 35 0.04240000 10.00
#> 36 0.15894737  0.95
#> 37 0.12000000  1.75
#> 38 0.15240000  2.50
#> 39 0.09714286  3.50
#> 40 0.07780000  5.00
#> 41 0.05866667  6.75
#> 42 0.04140000 10.00
#> 43 0.11157895  0.95
#> 44 0.10971429  1.75
#> 45 0.10480000  2.50
#> 46 0.08571429  3.50
#> 47 0.06180000  5.00
#> 48 0.04800000  6.75
#> 49 0.03550000 10.00
#> 50 0.12631579  0.95
#> 51 0.12571429  1.75
#> 52 0.12240000  2.50
#> 53 0.09085714  3.50
#> 54 0.06480000  5.00
#> 55 0.04607407  6.75
#> 56 0.03150000 10.00
#> 57 0.11894737  0.95
#> 58 0.11085714  1.75
#> 59 0.10320000  2.50
#> 60 0.07971429  3.50
#> 61 0.05700000  5.00
#> 62 0.04162963  6.75
#> 63 0.02780000 10.00
#> 64 0.11052632  0.95
#> 65 0.08514286  1.75
#> 66 0.07240000  2.50
#> 67 0.05400000  3.50
#> 68 0.03900000  5.00
#> 69 0.03288889  6.75
#> 70 0.02190000 10.00
#> 71 0.08105263  0.95
#> 72 0.06514286  1.75
#> 73 0.04920000  2.50
#> 74 0.03714286  3.50
#> 75 0.02500000  5.00
#> 76 0.02029630  6.75
#> 77 0.01440000 10.00
#> 78 0.11157895  0.95
#> 79 0.10285714  1.75
#> 80 0.07160000  2.50
#> 81 0.05114286  3.50
#> 82 0.03580000  5.00
#> 83 0.02800000  6.75
#> 84 0.01990000 10.00

  # summarise
  co2$summarise(mean=~mean(uptake),sd=~sd(uptake))
  co2
#>      mean       sd
#> 1 27.2131 10.81441
  
  co2 = m61r(CO2)
  co2$group_by(~cbind(Type,Treatment))
  co2$summarise(mean=~mean(uptake),sd=~sd(uptake))
  co2
#>          Type  Treatment     mean       sd
#> 1      Quebec nonchilled 35.33333 9.596371
#> 2 Mississippi nonchilled 25.95238 7.402136
#> 3      Quebec    chilled 31.75238 9.644823
#> 4 Mississippi    chilled 15.81429 4.058976

  # arrange/dessange
  co2$arrange(~c(conc))
  co2
#>    Plant        Type  Treatment conc uptake
#> 1    Qn1      Quebec nonchilled   95   16.0
#> 2    Qn2      Quebec nonchilled   95   13.6
#> 3    Qn3      Quebec nonchilled   95   16.2
#> 4    Qc1      Quebec    chilled   95   14.2
#> 5    Qc2      Quebec    chilled   95    9.3
#> 6    Qc3      Quebec    chilled   95   15.1
#> 7    Mn1 Mississippi nonchilled   95   10.6
#> 8    Mn2 Mississippi nonchilled   95   12.0
#> 9    Mn3 Mississippi nonchilled   95   11.3
#> 10   Mc1 Mississippi    chilled   95   10.5
#> 11   Mc2 Mississippi    chilled   95    7.7
#> 12   Mc3 Mississippi    chilled   95   10.6
#> 13   Qn1      Quebec nonchilled  175   30.4
#> 14   Qn2      Quebec nonchilled  175   27.3
#> 15   Qn3      Quebec nonchilled  175   32.4
#> 16   Qc1      Quebec    chilled  175   24.1
#> 17   Qc2      Quebec    chilled  175   27.3
#> 18   Qc3      Quebec    chilled  175   21.0
#> 19   Mn1 Mississippi nonchilled  175   19.2
#> 20   Mn2 Mississippi nonchilled  175   22.0
#> 21   Mn3 Mississippi nonchilled  175   19.4
#> 22   Mc1 Mississippi    chilled  175   14.9
#> 23   Mc2 Mississippi    chilled  175   11.4
#> 24   Mc3 Mississippi    chilled  175   18.0
#> 25   Qn1      Quebec nonchilled  250   34.8
#> 26   Qn2      Quebec nonchilled  250   37.1
#> 27   Qn3      Quebec nonchilled  250   40.3
#> 28   Qc1      Quebec    chilled  250   30.3
#> 29   Qc2      Quebec    chilled  250   35.0
#> 30   Qc3      Quebec    chilled  250   38.1
#> 31   Mn1 Mississippi nonchilled  250   26.2
#> 32   Mn2 Mississippi nonchilled  250   30.6
#> 33   Mn3 Mississippi nonchilled  250   25.8
#> 34   Mc1 Mississippi    chilled  250   18.1
#> 35   Mc2 Mississippi    chilled  250   12.3
#> 36   Mc3 Mississippi    chilled  250   17.9
#> 37   Qn1      Quebec nonchilled  350   37.2
#> 38   Qn2      Quebec nonchilled  350   41.8
#> 39   Qn3      Quebec nonchilled  350   42.1
#> 40   Qc1      Quebec    chilled  350   34.6
#> 41   Qc2      Quebec    chilled  350   38.8
#> 42   Qc3      Quebec    chilled  350   34.0
#> 43   Mn1 Mississippi nonchilled  350   30.0
#> 44   Mn2 Mississippi nonchilled  350   31.8
#> 45   Mn3 Mississippi nonchilled  350   27.9
#> 46   Mc1 Mississippi    chilled  350   18.9
#> 47   Mc2 Mississippi    chilled  350   13.0
#> 48   Mc3 Mississippi    chilled  350   17.9
#> 49   Qn1      Quebec nonchilled  500   35.3
#> 50   Qn2      Quebec nonchilled  500   40.6
#> 51   Qn3      Quebec nonchilled  500   42.9
#> 52   Qc1      Quebec    chilled  500   32.5
#> 53   Qc2      Quebec    chilled  500   38.6
#> 54   Qc3      Quebec    chilled  500   38.9
#> 55   Mn1 Mississippi nonchilled  500   30.9
#> 56   Mn2 Mississippi nonchilled  500   32.4
#> 57   Mn3 Mississippi nonchilled  500   28.5
#> 58   Mc1 Mississippi    chilled  500   19.5
#> 59   Mc2 Mississippi    chilled  500   12.5
#> 60   Mc3 Mississippi    chilled  500   17.9
#> 61   Qn1      Quebec nonchilled  675   39.2
#> 62   Qn2      Quebec nonchilled  675   41.4
#> 63   Qn3      Quebec nonchilled  675   43.9
#> 64   Qc1      Quebec    chilled  675   35.4
#> 65   Qc2      Quebec    chilled  675   37.5
#> 66   Qc3      Quebec    chilled  675   39.6
#> 67   Mn1 Mississippi nonchilled  675   32.4
#> 68   Mn2 Mississippi nonchilled  675   31.1
#> 69   Mn3 Mississippi nonchilled  675   28.1
#> 70   Mc1 Mississippi    chilled  675   22.2
#> 71   Mc2 Mississippi    chilled  675   13.7
#> 72   Mc3 Mississippi    chilled  675   18.9
#> 73   Qn1      Quebec nonchilled 1000   39.7
#> 74   Qn2      Quebec nonchilled 1000   44.3
#> 75   Qn3      Quebec nonchilled 1000   45.5
#> 76   Qc1      Quebec    chilled 1000   38.7
#> 77   Qc2      Quebec    chilled 1000   42.4
#> 78   Qc3      Quebec    chilled 1000   41.4
#> 79   Mn1 Mississippi nonchilled 1000   35.5
#> 80   Mn2 Mississippi nonchilled 1000   31.5
#> 81   Mn3 Mississippi nonchilled 1000   27.8
#> 82   Mc1 Mississippi    chilled 1000   21.9
#> 83   Mc2 Mississippi    chilled 1000   14.4
#> 84   Mc3 Mississippi    chilled 1000   19.9

  co2$arrange(~c(Treatment,conc,uptake))
  co2
#>    Plant        Type  Treatment conc uptake
#> 1    Mn1 Mississippi nonchilled   95   10.6
#> 2    Mn3 Mississippi nonchilled   95   11.3
#> 3    Mn2 Mississippi nonchilled   95   12.0
#> 4    Qn2      Quebec nonchilled   95   13.6
#> 5    Qn1      Quebec nonchilled   95   16.0
#> 6    Qn3      Quebec nonchilled   95   16.2
#> 7    Mn1 Mississippi nonchilled  175   19.2
#> 8    Mn3 Mississippi nonchilled  175   19.4
#> 9    Mn2 Mississippi nonchilled  175   22.0
#> 10   Qn2      Quebec nonchilled  175   27.3
#> 11   Qn1      Quebec nonchilled  175   30.4
#> 12   Qn3      Quebec nonchilled  175   32.4
#> 13   Mn3 Mississippi nonchilled  250   25.8
#> 14   Mn1 Mississippi nonchilled  250   26.2
#> 15   Mn2 Mississippi nonchilled  250   30.6
#> 16   Qn1      Quebec nonchilled  250   34.8
#> 17   Qn2      Quebec nonchilled  250   37.1
#> 18   Qn3      Quebec nonchilled  250   40.3
#> 19   Mn3 Mississippi nonchilled  350   27.9
#> 20   Mn1 Mississippi nonchilled  350   30.0
#> 21   Mn2 Mississippi nonchilled  350   31.8
#> 22   Qn1      Quebec nonchilled  350   37.2
#> 23   Qn2      Quebec nonchilled  350   41.8
#> 24   Qn3      Quebec nonchilled  350   42.1
#> 25   Mn3 Mississippi nonchilled  500   28.5
#> 26   Mn1 Mississippi nonchilled  500   30.9
#> 27   Mn2 Mississippi nonchilled  500   32.4
#> 28   Qn1      Quebec nonchilled  500   35.3
#> 29   Qn2      Quebec nonchilled  500   40.6
#> 30   Qn3      Quebec nonchilled  500   42.9
#> 31   Mn3 Mississippi nonchilled  675   28.1
#> 32   Mn2 Mississippi nonchilled  675   31.1
#> 33   Mn1 Mississippi nonchilled  675   32.4
#> 34   Qn1      Quebec nonchilled  675   39.2
#> 35   Qn2      Quebec nonchilled  675   41.4
#> 36   Qn3      Quebec nonchilled  675   43.9
#> 37   Mn3 Mississippi nonchilled 1000   27.8
#> 38   Mn2 Mississippi nonchilled 1000   31.5
#> 39   Mn1 Mississippi nonchilled 1000   35.5
#> 40   Qn1      Quebec nonchilled 1000   39.7
#> 41   Qn2      Quebec nonchilled 1000   44.3
#> 42   Qn3      Quebec nonchilled 1000   45.5
#> 43   Mc2 Mississippi    chilled   95    7.7
#> 44   Qc2      Quebec    chilled   95    9.3
#> 45   Mc1 Mississippi    chilled   95   10.5
#> 46   Mc3 Mississippi    chilled   95   10.6
#> 47   Qc1      Quebec    chilled   95   14.2
#> 48   Qc3      Quebec    chilled   95   15.1
#> 49   Mc2 Mississippi    chilled  175   11.4
#> 50   Mc1 Mississippi    chilled  175   14.9
#> 51   Mc3 Mississippi    chilled  175   18.0
#> 52   Qc3      Quebec    chilled  175   21.0
#> 53   Qc1      Quebec    chilled  175   24.1
#> 54   Qc2      Quebec    chilled  175   27.3
#> 55   Mc2 Mississippi    chilled  250   12.3
#> 56   Mc3 Mississippi    chilled  250   17.9
#> 57   Mc1 Mississippi    chilled  250   18.1
#> 58   Qc1      Quebec    chilled  250   30.3
#> 59   Qc2      Quebec    chilled  250   35.0
#> 60   Qc3      Quebec    chilled  250   38.1
#> 61   Mc2 Mississippi    chilled  350   13.0
#> 62   Mc3 Mississippi    chilled  350   17.9
#> 63   Mc1 Mississippi    chilled  350   18.9
#> 64   Qc3      Quebec    chilled  350   34.0
#> 65   Qc1      Quebec    chilled  350   34.6
#> 66   Qc2      Quebec    chilled  350   38.8
#> 67   Mc2 Mississippi    chilled  500   12.5
#> 68   Mc3 Mississippi    chilled  500   17.9
#> 69   Mc1 Mississippi    chilled  500   19.5
#> 70   Qc1      Quebec    chilled  500   32.5
#> 71   Qc2      Quebec    chilled  500   38.6
#> 72   Qc3      Quebec    chilled  500   38.9
#> 73   Mc2 Mississippi    chilled  675   13.7
#> 74   Mc3 Mississippi    chilled  675   18.9
#> 75   Mc1 Mississippi    chilled  675   22.2
#> 76   Qc1      Quebec    chilled  675   35.4
#> 77   Qc2      Quebec    chilled  675   37.5
#> 78   Qc3      Quebec    chilled  675   39.6
#> 79   Mc2 Mississippi    chilled 1000   14.4
#> 80   Mc3 Mississippi    chilled 1000   19.9
#> 81   Mc1 Mississippi    chilled 1000   21.9
#> 82   Qc1      Quebec    chilled 1000   38.7
#> 83   Qc3      Quebec    chilled 1000   41.4
#> 84   Qc2      Quebec    chilled 1000   42.4

  co2$desange(~c(Treatment,conc,uptake))
  co2
#>    Plant        Type  Treatment conc uptake
#> 1    Qc2      Quebec    chilled 1000   42.4
#> 2    Qc3      Quebec    chilled 1000   41.4
#> 3    Qc1      Quebec    chilled 1000   38.7
#> 4    Mc1 Mississippi    chilled 1000   21.9
#> 5    Mc3 Mississippi    chilled 1000   19.9
#> 6    Mc2 Mississippi    chilled 1000   14.4
#> 7    Qc3      Quebec    chilled  675   39.6
#> 8    Qc2      Quebec    chilled  675   37.5
#> 9    Qc1      Quebec    chilled  675   35.4
#> 10   Mc1 Mississippi    chilled  675   22.2
#> 11   Mc3 Mississippi    chilled  675   18.9
#> 12   Mc2 Mississippi    chilled  675   13.7
#> 13   Qc3      Quebec    chilled  500   38.9
#> 14   Qc2      Quebec    chilled  500   38.6
#> 15   Qc1      Quebec    chilled  500   32.5
#> 16   Mc1 Mississippi    chilled  500   19.5
#> 17   Mc3 Mississippi    chilled  500   17.9
#> 18   Mc2 Mississippi    chilled  500   12.5
#> 19   Qc2      Quebec    chilled  350   38.8
#> 20   Qc1      Quebec    chilled  350   34.6
#> 21   Qc3      Quebec    chilled  350   34.0
#> 22   Mc1 Mississippi    chilled  350   18.9
#> 23   Mc3 Mississippi    chilled  350   17.9
#> 24   Mc2 Mississippi    chilled  350   13.0
#> 25   Qc3      Quebec    chilled  250   38.1
#> 26   Qc2      Quebec    chilled  250   35.0
#> 27   Qc1      Quebec    chilled  250   30.3
#> 28   Mc1 Mississippi    chilled  250   18.1
#> 29   Mc3 Mississippi    chilled  250   17.9
#> 30   Mc2 Mississippi    chilled  250   12.3
#> 31   Qc2      Quebec    chilled  175   27.3
#> 32   Qc1      Quebec    chilled  175   24.1
#> 33   Qc3      Quebec    chilled  175   21.0
#> 34   Mc3 Mississippi    chilled  175   18.0
#> 35   Mc1 Mississippi    chilled  175   14.9
#> 36   Mc2 Mississippi    chilled  175   11.4
#> 37   Qc3      Quebec    chilled   95   15.1
#> 38   Qc1      Quebec    chilled   95   14.2
#> 39   Mc3 Mississippi    chilled   95   10.6
#> 40   Mc1 Mississippi    chilled   95   10.5
#> 41   Qc2      Quebec    chilled   95    9.3
#> 42   Mc2 Mississippi    chilled   95    7.7
#> 43   Qn3      Quebec nonchilled 1000   45.5
#> 44   Qn2      Quebec nonchilled 1000   44.3
#> 45   Qn1      Quebec nonchilled 1000   39.7
#> 46   Mn1 Mississippi nonchilled 1000   35.5
#> 47   Mn2 Mississippi nonchilled 1000   31.5
#> 48   Mn3 Mississippi nonchilled 1000   27.8
#> 49   Qn3      Quebec nonchilled  675   43.9
#> 50   Qn2      Quebec nonchilled  675   41.4
#> 51   Qn1      Quebec nonchilled  675   39.2
#> 52   Mn1 Mississippi nonchilled  675   32.4
#> 53   Mn2 Mississippi nonchilled  675   31.1
#> 54   Mn3 Mississippi nonchilled  675   28.1
#> 55   Qn3      Quebec nonchilled  500   42.9
#> 56   Qn2      Quebec nonchilled  500   40.6
#> 57   Qn1      Quebec nonchilled  500   35.3
#> 58   Mn2 Mississippi nonchilled  500   32.4
#> 59   Mn1 Mississippi nonchilled  500   30.9
#> 60   Mn3 Mississippi nonchilled  500   28.5
#> 61   Qn3      Quebec nonchilled  350   42.1
#> 62   Qn2      Quebec nonchilled  350   41.8
#> 63   Qn1      Quebec nonchilled  350   37.2
#> 64   Mn2 Mississippi nonchilled  350   31.8
#> 65   Mn1 Mississippi nonchilled  350   30.0
#> 66   Mn3 Mississippi nonchilled  350   27.9
#> 67   Qn3      Quebec nonchilled  250   40.3
#> 68   Qn2      Quebec nonchilled  250   37.1
#> 69   Qn1      Quebec nonchilled  250   34.8
#> 70   Mn2 Mississippi nonchilled  250   30.6
#> 71   Mn1 Mississippi nonchilled  250   26.2
#> 72   Mn3 Mississippi nonchilled  250   25.8
#> 73   Qn3      Quebec nonchilled  175   32.4
#> 74   Qn1      Quebec nonchilled  175   30.4
#> 75   Qn2      Quebec nonchilled  175   27.3
#> 76   Mn2 Mississippi nonchilled  175   22.0
#> 77   Mn3 Mississippi nonchilled  175   19.4
#> 78   Mn1 Mississippi nonchilled  175   19.2
#> 79   Qn3      Quebec nonchilled   95   16.2
#> 80   Qn1      Quebec nonchilled   95   16.0
#> 81   Qn2      Quebec nonchilled   95   13.6
#> 82   Mn2 Mississippi nonchilled   95   12.0
#> 83   Mn3 Mississippi nonchilled   95   11.3
#> 84   Mn1 Mississippi nonchilled   95   10.6

  # Reshape

  ## gather
  df3 <- data.frame(id = 1:4,
                    age = c(40,50,60,50),
                    dose.a1 = c(1,2,1,2),
                    dose.a2 = c(2,1,2,1),
                    dose.a14 = c(3,3,3,3))

  df4 <- m61r::m61r(df3)
  df4$gather(pivot = c("id","age"))
  df4
#>    id age parameters values
#> 1   1  40    dose.a1      1
#> 2   2  50    dose.a1      2
#> 3   3  60    dose.a1      1
#> 4   4  50    dose.a1      2
#> 5   1  40    dose.a2      2
#> 6   2  50    dose.a2      1
#> 7   3  60    dose.a2      2
#> 8   4  50    dose.a2      1
#> 9   1  40   dose.a14      3
#> 10  2  50   dose.a14      3
#> 11  3  60   dose.a14      3
#> 12  4  50   dose.a14      3

  ## spread
  df3 <- data.frame(id = 1:4,
                    age = c(40,50,60,50),
                    dose.a1 = c(1,2,1,2),
                    dose.a2 = c(2,1,2,1),
                    dose.a14 = c(3,3,3,3))

  df4 <- m61r::gather_(df3,pivot = c("id","age"))
  df4 <- rbind(df4,
    data.frame(id=5, age=20,parameters="dose.a14",values=8),
    data.frame(id=6, age=10,parameters="dose.a1",values=5))

  tmp <- m61r::m61r(df4)
  tmp$spread(col_name="parameters",col_values="values",pivot=c("id","age"))
  tmp
#>    id age dose.a1 dose.a2 dose.a14
#> 1   1  40       1       2        3
#> 2   2  50       2       1        3
#> 3   3  60       1       2        3
#> 4   4  50       2       1        3
#> 13  5  20      NA      NA        8
#> 14  6  10       5      NA       NA

  # equivalence
  co2           # is not equivalent to co2[]
#>    Plant        Type  Treatment conc uptake
#> 1    Qn1      Quebec nonchilled   95   16.0
#> 2    Qn1      Quebec nonchilled  175   30.4
#> 3    Qn1      Quebec nonchilled  250   34.8
#> 4    Qn1      Quebec nonchilled  350   37.2
#> 5    Qn1      Quebec nonchilled  500   35.3
#> 6    Qn1      Quebec nonchilled  675   39.2
#> 7    Qn1      Quebec nonchilled 1000   39.7
#> 8    Qn2      Quebec nonchilled   95   13.6
#> 9    Qn2      Quebec nonchilled  175   27.3
#> 10   Qn2      Quebec nonchilled  250   37.1
#> 11   Qn2      Quebec nonchilled  350   41.8
#> 12   Qn2      Quebec nonchilled  500   40.6
#> 13   Qn2      Quebec nonchilled  675   41.4
#> 14   Qn2      Quebec nonchilled 1000   44.3
#> 15   Qn3      Quebec nonchilled   95   16.2
#> 16   Qn3      Quebec nonchilled  175   32.4
#> 17   Qn3      Quebec nonchilled  250   40.3
#> 18   Qn3      Quebec nonchilled  350   42.1
#> 19   Qn3      Quebec nonchilled  500   42.9
#> 20   Qn3      Quebec nonchilled  675   43.9
#> 21   Qn3      Quebec nonchilled 1000   45.5
#> 22   Qc1      Quebec    chilled   95   14.2
#> 23   Qc1      Quebec    chilled  175   24.1
#> 24   Qc1      Quebec    chilled  250   30.3
#> 25   Qc1      Quebec    chilled  350   34.6
#> 26   Qc1      Quebec    chilled  500   32.5
#> 27   Qc1      Quebec    chilled  675   35.4
#> 28   Qc1      Quebec    chilled 1000   38.7
#> 29   Qc2      Quebec    chilled   95    9.3
#> 30   Qc2      Quebec    chilled  175   27.3
#> 31   Qc2      Quebec    chilled  250   35.0
#> 32   Qc2      Quebec    chilled  350   38.8
#> 33   Qc2      Quebec    chilled  500   38.6
#> 34   Qc2      Quebec    chilled  675   37.5
#> 35   Qc2      Quebec    chilled 1000   42.4
#> 36   Qc3      Quebec    chilled   95   15.1
#> 37   Qc3      Quebec    chilled  175   21.0
#> 38   Qc3      Quebec    chilled  250   38.1
#> 39   Qc3      Quebec    chilled  350   34.0
#> 40   Qc3      Quebec    chilled  500   38.9
#> 41   Qc3      Quebec    chilled  675   39.6
#> 42   Qc3      Quebec    chilled 1000   41.4
#> 43   Mn1 Mississippi nonchilled   95   10.6
#> 44   Mn1 Mississippi nonchilled  175   19.2
#> 45   Mn1 Mississippi nonchilled  250   26.2
#> 46   Mn1 Mississippi nonchilled  350   30.0
#> 47   Mn1 Mississippi nonchilled  500   30.9
#> 48   Mn1 Mississippi nonchilled  675   32.4
#> 49   Mn1 Mississippi nonchilled 1000   35.5
#> 50   Mn2 Mississippi nonchilled   95   12.0
#> 51   Mn2 Mississippi nonchilled  175   22.0
#> 52   Mn2 Mississippi nonchilled  250   30.6
#> 53   Mn2 Mississippi nonchilled  350   31.8
#> 54   Mn2 Mississippi nonchilled  500   32.4
#> 55   Mn2 Mississippi nonchilled  675   31.1
#> 56   Mn2 Mississippi nonchilled 1000   31.5
#> 57   Mn3 Mississippi nonchilled   95   11.3
#> 58   Mn3 Mississippi nonchilled  175   19.4
#> 59   Mn3 Mississippi nonchilled  250   25.8
#> 60   Mn3 Mississippi nonchilled  350   27.9
#> 61   Mn3 Mississippi nonchilled  500   28.5
#> 62   Mn3 Mississippi nonchilled  675   28.1
#> 63   Mn3 Mississippi nonchilled 1000   27.8
#> 64   Mc1 Mississippi    chilled   95   10.5
#> 65   Mc1 Mississippi    chilled  175   14.9
#> 66   Mc1 Mississippi    chilled  250   18.1
#> 67   Mc1 Mississippi    chilled  350   18.9
#> 68   Mc1 Mississippi    chilled  500   19.5
#> 69   Mc1 Mississippi    chilled  675   22.2
#> 70   Mc1 Mississippi    chilled 1000   21.9
#> 71   Mc2 Mississippi    chilled   95    7.7
#> 72   Mc2 Mississippi    chilled  175   11.4
#> 73   Mc2 Mississippi    chilled  250   12.3
#> 74   Mc2 Mississippi    chilled  350   13.0
#> 75   Mc2 Mississippi    chilled  500   12.5
#> 76   Mc2 Mississippi    chilled  675   13.7
#> 77   Mc2 Mississippi    chilled 1000   14.4
#> 78   Mc3 Mississippi    chilled   95   10.6
#> 79   Mc3 Mississippi    chilled  175   18.0
#> 80   Mc3 Mississippi    chilled  250   17.9
#> 81   Mc3 Mississippi    chilled  350   17.9
#> 82   Mc3 Mississippi    chilled  500   17.9
#> 83   Mc3 Mississippi    chilled  675   18.9
#> 84   Mc3 Mississippi    chilled 1000   19.9
  co2[]         # is equivalent to co2$values()
#>    Plant        Type  Treatment conc uptake
#> 1    Qn1      Quebec nonchilled   95   16.0
#> 2    Qn1      Quebec nonchilled  175   30.4
#> 3    Qn1      Quebec nonchilled  250   34.8
#> 4    Qn1      Quebec nonchilled  350   37.2
#> 5    Qn1      Quebec nonchilled  500   35.3
#> 6    Qn1      Quebec nonchilled  675   39.2
#> 7    Qn1      Quebec nonchilled 1000   39.7
#> 8    Qn2      Quebec nonchilled   95   13.6
#> 9    Qn2      Quebec nonchilled  175   27.3
#> 10   Qn2      Quebec nonchilled  250   37.1
#> 11   Qn2      Quebec nonchilled  350   41.8
#> 12   Qn2      Quebec nonchilled  500   40.6
#> 13   Qn2      Quebec nonchilled  675   41.4
#> 14   Qn2      Quebec nonchilled 1000   44.3
#> 15   Qn3      Quebec nonchilled   95   16.2
#> 16   Qn3      Quebec nonchilled  175   32.4
#> 17   Qn3      Quebec nonchilled  250   40.3
#> 18   Qn3      Quebec nonchilled  350   42.1
#> 19   Qn3      Quebec nonchilled  500   42.9
#> 20   Qn3      Quebec nonchilled  675   43.9
#> 21   Qn3      Quebec nonchilled 1000   45.5
#> 22   Qc1      Quebec    chilled   95   14.2
#> 23   Qc1      Quebec    chilled  175   24.1
#> 24   Qc1      Quebec    chilled  250   30.3
#> 25   Qc1      Quebec    chilled  350   34.6
#> 26   Qc1      Quebec    chilled  500   32.5
#> 27   Qc1      Quebec    chilled  675   35.4
#> 28   Qc1      Quebec    chilled 1000   38.7
#> 29   Qc2      Quebec    chilled   95    9.3
#> 30   Qc2      Quebec    chilled  175   27.3
#> 31   Qc2      Quebec    chilled  250   35.0
#> 32   Qc2      Quebec    chilled  350   38.8
#> 33   Qc2      Quebec    chilled  500   38.6
#> 34   Qc2      Quebec    chilled  675   37.5
#> 35   Qc2      Quebec    chilled 1000   42.4
#> 36   Qc3      Quebec    chilled   95   15.1
#> 37   Qc3      Quebec    chilled  175   21.0
#> 38   Qc3      Quebec    chilled  250   38.1
#> 39   Qc3      Quebec    chilled  350   34.0
#> 40   Qc3      Quebec    chilled  500   38.9
#> 41   Qc3      Quebec    chilled  675   39.6
#> 42   Qc3      Quebec    chilled 1000   41.4
#> 43   Mn1 Mississippi nonchilled   95   10.6
#> 44   Mn1 Mississippi nonchilled  175   19.2
#> 45   Mn1 Mississippi nonchilled  250   26.2
#> 46   Mn1 Mississippi nonchilled  350   30.0
#> 47   Mn1 Mississippi nonchilled  500   30.9
#> 48   Mn1 Mississippi nonchilled  675   32.4
#> 49   Mn1 Mississippi nonchilled 1000   35.5
#> 50   Mn2 Mississippi nonchilled   95   12.0
#> 51   Mn2 Mississippi nonchilled  175   22.0
#> 52   Mn2 Mississippi nonchilled  250   30.6
#> 53   Mn2 Mississippi nonchilled  350   31.8
#> 54   Mn2 Mississippi nonchilled  500   32.4
#> 55   Mn2 Mississippi nonchilled  675   31.1
#> 56   Mn2 Mississippi nonchilled 1000   31.5
#> 57   Mn3 Mississippi nonchilled   95   11.3
#> 58   Mn3 Mississippi nonchilled  175   19.4
#> 59   Mn3 Mississippi nonchilled  250   25.8
#> 60   Mn3 Mississippi nonchilled  350   27.9
#> 61   Mn3 Mississippi nonchilled  500   28.5
#> 62   Mn3 Mississippi nonchilled  675   28.1
#> 63   Mn3 Mississippi nonchilled 1000   27.8
#> 64   Mc1 Mississippi    chilled   95   10.5
#> 65   Mc1 Mississippi    chilled  175   14.9
#> 66   Mc1 Mississippi    chilled  250   18.1
#> 67   Mc1 Mississippi    chilled  350   18.9
#> 68   Mc1 Mississippi    chilled  500   19.5
#> 69   Mc1 Mississippi    chilled  675   22.2
#> 70   Mc1 Mississippi    chilled 1000   21.9
#> 71   Mc2 Mississippi    chilled   95    7.7
#> 72   Mc2 Mississippi    chilled  175   11.4
#> 73   Mc2 Mississippi    chilled  250   12.3
#> 74   Mc2 Mississippi    chilled  350   13.0
#> 75   Mc2 Mississippi    chilled  500   12.5
#> 76   Mc2 Mississippi    chilled  675   13.7
#> 77   Mc2 Mississippi    chilled 1000   14.4
#> 78   Mc3 Mississippi    chilled   95   10.6
#> 79   Mc3 Mississippi    chilled  175   18.0
#> 80   Mc3 Mississippi    chilled  250   17.9
#> 81   Mc3 Mississippi    chilled  350   17.9
#> 82   Mc3 Mississippi    chilled  500   17.9
#> 83   Mc3 Mississippi    chilled  675   18.9
#> 84   Mc3 Mississippi    chilled 1000   19.9
  co2[1,]       # is equivalent to co2$values(1,)
#>   Plant   Type  Treatment conc uptake
#> 1   Qn1 Quebec nonchilled   95     16
  co2[,2:3]     # is equivalent to co2$values(,2:3)
#>           Type  Treatment
#> 1       Quebec nonchilled
#> 2       Quebec nonchilled
#> 3       Quebec nonchilled
#> 4       Quebec nonchilled
#> 5       Quebec nonchilled
#> 6       Quebec nonchilled
#> 7       Quebec nonchilled
#> 8       Quebec nonchilled
#> 9       Quebec nonchilled
#> 10      Quebec nonchilled
#> 11      Quebec nonchilled
#> 12      Quebec nonchilled
#> 13      Quebec nonchilled
#> 14      Quebec nonchilled
#> 15      Quebec nonchilled
#> 16      Quebec nonchilled
#> 17      Quebec nonchilled
#> 18      Quebec nonchilled
#> 19      Quebec nonchilled
#> 20      Quebec nonchilled
#> 21      Quebec nonchilled
#> 22      Quebec    chilled
#> 23      Quebec    chilled
#> 24      Quebec    chilled
#> 25      Quebec    chilled
#> 26      Quebec    chilled
#> 27      Quebec    chilled
#> 28      Quebec    chilled
#> 29      Quebec    chilled
#> 30      Quebec    chilled
#> 31      Quebec    chilled
#> 32      Quebec    chilled
#> 33      Quebec    chilled
#> 34      Quebec    chilled
#> 35      Quebec    chilled
#> 36      Quebec    chilled
#> 37      Quebec    chilled
#> 38      Quebec    chilled
#> 39      Quebec    chilled
#> 40      Quebec    chilled
#> 41      Quebec    chilled
#> 42      Quebec    chilled
#> 43 Mississippi nonchilled
#> 44 Mississippi nonchilled
#> 45 Mississippi nonchilled
#> 46 Mississippi nonchilled
#> 47 Mississippi nonchilled
#> 48 Mississippi nonchilled
#> 49 Mississippi nonchilled
#> 50 Mississippi nonchilled
#> 51 Mississippi nonchilled
#> 52 Mississippi nonchilled
#> 53 Mississippi nonchilled
#> 54 Mississippi nonchilled
#> 55 Mississippi nonchilled
#> 56 Mississippi nonchilled
#> 57 Mississippi nonchilled
#> 58 Mississippi nonchilled
#> 59 Mississippi nonchilled
#> 60 Mississippi nonchilled
#> 61 Mississippi nonchilled
#> 62 Mississippi nonchilled
#> 63 Mississippi nonchilled
#> 64 Mississippi    chilled
#> 65 Mississippi    chilled
#> 66 Mississippi    chilled
#> 67 Mississippi    chilled
#> 68 Mississippi    chilled
#> 69 Mississippi    chilled
#> 70 Mississippi    chilled
#> 71 Mississippi    chilled
#> 72 Mississippi    chilled
#> 73 Mississippi    chilled
#> 74 Mississippi    chilled
#> 75 Mississippi    chilled
#> 76 Mississippi    chilled
#> 77 Mississippi    chilled
#> 78 Mississippi    chilled
#> 79 Mississippi    chilled
#> 80 Mississippi    chilled
#> 81 Mississippi    chilled
#> 82 Mississippi    chilled
#> 83 Mississippi    chilled
#> 84 Mississippi    chilled
  co2[1:10,1:3] # is equivalent to co2$values(1:10,2:3)
#>    Plant   Type  Treatment
#> 1    Qn1 Quebec nonchilled
#> 2    Qn1 Quebec nonchilled
#> 3    Qn1 Quebec nonchilled
#> 4    Qn1 Quebec nonchilled
#> 5    Qn1 Quebec nonchilled
#> 6    Qn1 Quebec nonchilled
#> 7    Qn1 Quebec nonchilled
#> 8    Qn2 Quebec nonchilled
#> 9    Qn2 Quebec nonchilled
#> 10   Qn2 Quebec nonchilled
  co2[1,"Plant"]# is equivalent to co2$values(1,"Plant")
#>   Plant
#> 1   Qn1

  # modification on m61r object only stay for one step
  co2[1,"conc"] <- 100
  co2[1,] # temporary result
#>   Plant   Type  Treatment conc uptake
#> 1   Qn1 Quebec nonchilled  100     16
  co2[1,] # back to normal
#>   Plant   Type  Treatment conc uptake
#> 1   Qn1 Quebec nonchilled   95     16

  # WARNING:Keep the brackets to manipulate the intern data.frame
  # ... OR you will destroy co2, and only keep the data.frame
  # co2 <- co2[-1,]
  # class(co2) # data.frame

  # descriptive manipulation
  names(co2)
#> [1] "Plant"     "Type"      "Treatment" "conc"      "uptake"   
  dim(co2)
#> [1] 84  5
  str(co2)
#> Classes 'm61r', 'environment' <environment: 0x55a6b861c120> 

  ## cloning
  # The following will only create a second variable that point on
  # the same object (!= cloning)
  foo <- co2
  str(co2)
#> Classes 'm61r', 'environment' <environment: 0x55a6b861c120> 
  str(foo)
#> Classes 'm61r', 'environment' <environment: 0x55a6b861c120> 

  # Instead, cloning into a new environemnt
  foo <- co2$clone()
  str(co2)
#> Classes 'm61r', 'environment' <environment: 0x55a6b861c120> 
  str(foo)
#> Classes 'm61r', 'environment' <environment: 0x55a6c3c42660>