Mutate and transmutate a data.frame

Mutate and transmutate a data.frame.

Usage

# mutate_(df, ...)
# transmutate_(df, ...)

Details

mutate_ and transmutate_ execute expressions non-grouped. If the m61r object is in a grouped state (via $group_by()), that grouping state is ignored by the primitive functions, ensuring Base R speed for vectorised operations.

Arguments

df

data.frame

...

formula used for mutating/transmutating the data.frame

Value

The functions return a data frame. The output has the following properties:

For function mutate_(), output includes all df columns. In addition, new columns are created according to argument ... and placed after the others.

For function transmutate_(), output includes only columns created according to argument ... and placed after the others.

Examples

tmp <- mutate_(CO2,z=~conc/uptake)
head(tmp)
#>   Plant   Type  Treatment conc uptake         z
#> 1   Qn1 Quebec nonchilled   95   16.0  5.937500
#> 2   Qn1 Quebec nonchilled  175   30.4  5.756579
#> 3   Qn1 Quebec nonchilled  250   34.8  7.183908
#> 4   Qn1 Quebec nonchilled  350   37.2  9.408602
#> 5   Qn1 Quebec nonchilled  500   35.3 14.164306
#> 6   Qn1 Quebec nonchilled  675   39.2 17.219388

# Return an warning: expression mean(uptake) get a result with 'nrow' different from 'df'
# tmp <- mutate_(CO2,mean=~mean(uptake))

tmp <- mutate_(CO2,z1=~uptake/conc,y=~conc/100)
head(tmp)
#>   Plant   Type  Treatment conc uptake         z1    y
#> 1   Qn1 Quebec nonchilled   95   16.0 0.16842105 0.95
#> 2   Qn1 Quebec nonchilled  175   30.4 0.17371429 1.75
#> 3   Qn1 Quebec nonchilled  250   34.8 0.13920000 2.50
#> 4   Qn1 Quebec nonchilled  350   37.2 0.10628571 3.50
#> 5   Qn1 Quebec nonchilled  500   35.3 0.07060000 5.00
#> 6   Qn1 Quebec nonchilled  675   39.2 0.05807407 6.75

tmp <- transmutate_(CO2,z2=~uptake/conc,y2=~conc/100)
head(tmp)
#>           z2   y2
#> 1 0.16842105 0.95
#> 2 0.17371429 1.75
#> 3 0.13920000 2.50
#> 4 0.10628571 3.50
#> 5 0.07060000 5.00
#> 6 0.05807407 6.75

# with m61r class
co2 <- m61r(df=CO2)

co2$mutate(z=~conc/uptake)
co2
#>    Plant        Type  Treatment conc uptake         z
#> 1    Qn1      Quebec nonchilled   95   16.0  5.937500
#> 2    Qn1      Quebec nonchilled  175   30.4  5.756579
#> 3    Qn1      Quebec nonchilled  250   34.8  7.183908
#> 4    Qn1      Quebec nonchilled  350   37.2  9.408602
#> 5    Qn1      Quebec nonchilled  500   35.3 14.164306
#> 6    Qn1      Quebec nonchilled  675   39.2 17.219388
#> 7    Qn1      Quebec nonchilled 1000   39.7 25.188917
#> 8    Qn2      Quebec nonchilled   95   13.6  6.985294
#> 9    Qn2      Quebec nonchilled  175   27.3  6.410256
#> 10   Qn2      Quebec nonchilled  250   37.1  6.738544
#> 11   Qn2      Quebec nonchilled  350   41.8  8.373206
#> 12   Qn2      Quebec nonchilled  500   40.6 12.315271
#> 13   Qn2      Quebec nonchilled  675   41.4 16.304348
#> 14   Qn2      Quebec nonchilled 1000   44.3 22.573363
#> 15   Qn3      Quebec nonchilled   95   16.2  5.864198
#> 16   Qn3      Quebec nonchilled  175   32.4  5.401235
#> 17   Qn3      Quebec nonchilled  250   40.3  6.203474
#> 18   Qn3      Quebec nonchilled  350   42.1  8.313539
#> 19   Qn3      Quebec nonchilled  500   42.9 11.655012
#> 20   Qn3      Quebec nonchilled  675   43.9 15.375854
#> 21   Qn3      Quebec nonchilled 1000   45.5 21.978022
#> 22   Qc1      Quebec    chilled   95   14.2  6.690141
#> 23   Qc1      Quebec    chilled  175   24.1  7.261411
#> 24   Qc1      Quebec    chilled  250   30.3  8.250825
#> 25   Qc1      Quebec    chilled  350   34.6 10.115607
#> 26   Qc1      Quebec    chilled  500   32.5 15.384615
#> 27   Qc1      Quebec    chilled  675   35.4 19.067797
#> 28   Qc1      Quebec    chilled 1000   38.7 25.839793
#> 29   Qc2      Quebec    chilled   95    9.3 10.215054
#> 30   Qc2      Quebec    chilled  175   27.3  6.410256
#> 31   Qc2      Quebec    chilled  250   35.0  7.142857
#> 32   Qc2      Quebec    chilled  350   38.8  9.020619
#> 33   Qc2      Quebec    chilled  500   38.6 12.953368
#> 34   Qc2      Quebec    chilled  675   37.5 18.000000
#> 35   Qc2      Quebec    chilled 1000   42.4 23.584906
#> 36   Qc3      Quebec    chilled   95   15.1  6.291391
#> 37   Qc3      Quebec    chilled  175   21.0  8.333333
#> 38   Qc3      Quebec    chilled  250   38.1  6.561680
#> 39   Qc3      Quebec    chilled  350   34.0 10.294118
#> 40   Qc3      Quebec    chilled  500   38.9 12.853470
#> 41   Qc3      Quebec    chilled  675   39.6 17.045455
#> 42   Qc3      Quebec    chilled 1000   41.4 24.154589
#> 43   Mn1 Mississippi nonchilled   95   10.6  8.962264
#> 44   Mn1 Mississippi nonchilled  175   19.2  9.114583
#> 45   Mn1 Mississippi nonchilled  250   26.2  9.541985
#> 46   Mn1 Mississippi nonchilled  350   30.0 11.666667
#> 47   Mn1 Mississippi nonchilled  500   30.9 16.181230
#> 48   Mn1 Mississippi nonchilled  675   32.4 20.833333
#> 49   Mn1 Mississippi nonchilled 1000   35.5 28.169014
#> 50   Mn2 Mississippi nonchilled   95   12.0  7.916667
#> 51   Mn2 Mississippi nonchilled  175   22.0  7.954545
#> 52   Mn2 Mississippi nonchilled  250   30.6  8.169935
#> 53   Mn2 Mississippi nonchilled  350   31.8 11.006289
#> 54   Mn2 Mississippi nonchilled  500   32.4 15.432099
#> 55   Mn2 Mississippi nonchilled  675   31.1 21.704180
#> 56   Mn2 Mississippi nonchilled 1000   31.5 31.746032
#> 57   Mn3 Mississippi nonchilled   95   11.3  8.407080
#> 58   Mn3 Mississippi nonchilled  175   19.4  9.020619
#> 59   Mn3 Mississippi nonchilled  250   25.8  9.689922
#> 60   Mn3 Mississippi nonchilled  350   27.9 12.544803
#> 61   Mn3 Mississippi nonchilled  500   28.5 17.543860
#> 62   Mn3 Mississippi nonchilled  675   28.1 24.021352
#> 63   Mn3 Mississippi nonchilled 1000   27.8 35.971223
#> 64   Mc1 Mississippi    chilled   95   10.5  9.047619
#> 65   Mc1 Mississippi    chilled  175   14.9 11.744966
#> 66   Mc1 Mississippi    chilled  250   18.1 13.812155
#> 67   Mc1 Mississippi    chilled  350   18.9 18.518519
#> 68   Mc1 Mississippi    chilled  500   19.5 25.641026
#> 69   Mc1 Mississippi    chilled  675   22.2 30.405405
#> 70   Mc1 Mississippi    chilled 1000   21.9 45.662100
#> 71   Mc2 Mississippi    chilled   95    7.7 12.337662
#> 72   Mc2 Mississippi    chilled  175   11.4 15.350877
#> 73   Mc2 Mississippi    chilled  250   12.3 20.325203
#> 74   Mc2 Mississippi    chilled  350   13.0 26.923077
#> 75   Mc2 Mississippi    chilled  500   12.5 40.000000
#> 76   Mc2 Mississippi    chilled  675   13.7 49.270073
#> 77   Mc2 Mississippi    chilled 1000   14.4 69.444444
#> 78   Mc3 Mississippi    chilled   95   10.6  8.962264
#> 79   Mc3 Mississippi    chilled  175   18.0  9.722222
#> 80   Mc3 Mississippi    chilled  250   17.9 13.966480
#> 81   Mc3 Mississippi    chilled  350   17.9 19.553073
#> 82   Mc3 Mississippi    chilled  500   17.9 27.932961
#> 83   Mc3 Mississippi    chilled  675   18.9 35.714286
#> 84   Mc3 Mississippi    chilled 1000   19.9 50.251256

# not allowed
#co2$mutate(mean=~mean(uptake))
#co2
#Warning message:
#In FUN(X[[i]], ...) : Expression mean has incompatible length.

co2$mutate(z1=~uptake/conc,y=~conc/100)
co2
#>    Plant        Type  Treatment conc uptake         z1     y
#> 1    Qn1      Quebec nonchilled   95   16.0 0.16842105  0.95
#> 2    Qn1      Quebec nonchilled  175   30.4 0.17371429  1.75
#> 3    Qn1      Quebec nonchilled  250   34.8 0.13920000  2.50
#> 4    Qn1      Quebec nonchilled  350   37.2 0.10628571  3.50
#> 5    Qn1      Quebec nonchilled  500   35.3 0.07060000  5.00
#> 6    Qn1      Quebec nonchilled  675   39.2 0.05807407  6.75
#> 7    Qn1      Quebec nonchilled 1000   39.7 0.03970000 10.00
#> 8    Qn2      Quebec nonchilled   95   13.6 0.14315789  0.95
#> 9    Qn2      Quebec nonchilled  175   27.3 0.15600000  1.75
#> 10   Qn2      Quebec nonchilled  250   37.1 0.14840000  2.50
#> 11   Qn2      Quebec nonchilled  350   41.8 0.11942857  3.50
#> 12   Qn2      Quebec nonchilled  500   40.6 0.08120000  5.00
#> 13   Qn2      Quebec nonchilled  675   41.4 0.06133333  6.75
#> 14   Qn2      Quebec nonchilled 1000   44.3 0.04430000 10.00
#> 15   Qn3      Quebec nonchilled   95   16.2 0.17052632  0.95
#> 16   Qn3      Quebec nonchilled  175   32.4 0.18514286  1.75
#> 17   Qn3      Quebec nonchilled  250   40.3 0.16120000  2.50
#> 18   Qn3      Quebec nonchilled  350   42.1 0.12028571  3.50
#> 19   Qn3      Quebec nonchilled  500   42.9 0.08580000  5.00
#> 20   Qn3      Quebec nonchilled  675   43.9 0.06503704  6.75
#> 21   Qn3      Quebec nonchilled 1000   45.5 0.04550000 10.00
#> 22   Qc1      Quebec    chilled   95   14.2 0.14947368  0.95
#> 23   Qc1      Quebec    chilled  175   24.1 0.13771429  1.75
#> 24   Qc1      Quebec    chilled  250   30.3 0.12120000  2.50
#> 25   Qc1      Quebec    chilled  350   34.6 0.09885714  3.50
#> 26   Qc1      Quebec    chilled  500   32.5 0.06500000  5.00
#> 27   Qc1      Quebec    chilled  675   35.4 0.05244444  6.75
#> 28   Qc1      Quebec    chilled 1000   38.7 0.03870000 10.00
#> 29   Qc2      Quebec    chilled   95    9.3 0.09789474  0.95
#> 30   Qc2      Quebec    chilled  175   27.3 0.15600000  1.75
#> 31   Qc2      Quebec    chilled  250   35.0 0.14000000  2.50
#> 32   Qc2      Quebec    chilled  350   38.8 0.11085714  3.50
#> 33   Qc2      Quebec    chilled  500   38.6 0.07720000  5.00
#> 34   Qc2      Quebec    chilled  675   37.5 0.05555556  6.75
#> 35   Qc2      Quebec    chilled 1000   42.4 0.04240000 10.00
#> 36   Qc3      Quebec    chilled   95   15.1 0.15894737  0.95
#> 37   Qc3      Quebec    chilled  175   21.0 0.12000000  1.75
#> 38   Qc3      Quebec    chilled  250   38.1 0.15240000  2.50
#> 39   Qc3      Quebec    chilled  350   34.0 0.09714286  3.50
#> 40   Qc3      Quebec    chilled  500   38.9 0.07780000  5.00
#> 41   Qc3      Quebec    chilled  675   39.6 0.05866667  6.75
#> 42   Qc3      Quebec    chilled 1000   41.4 0.04140000 10.00
#> 43   Mn1 Mississippi nonchilled   95   10.6 0.11157895  0.95
#> 44   Mn1 Mississippi nonchilled  175   19.2 0.10971429  1.75
#> 45   Mn1 Mississippi nonchilled  250   26.2 0.10480000  2.50
#> 46   Mn1 Mississippi nonchilled  350   30.0 0.08571429  3.50
#> 47   Mn1 Mississippi nonchilled  500   30.9 0.06180000  5.00
#> 48   Mn1 Mississippi nonchilled  675   32.4 0.04800000  6.75
#> 49   Mn1 Mississippi nonchilled 1000   35.5 0.03550000 10.00
#> 50   Mn2 Mississippi nonchilled   95   12.0 0.12631579  0.95
#> 51   Mn2 Mississippi nonchilled  175   22.0 0.12571429  1.75
#> 52   Mn2 Mississippi nonchilled  250   30.6 0.12240000  2.50
#> 53   Mn2 Mississippi nonchilled  350   31.8 0.09085714  3.50
#> 54   Mn2 Mississippi nonchilled  500   32.4 0.06480000  5.00
#> 55   Mn2 Mississippi nonchilled  675   31.1 0.04607407  6.75
#> 56   Mn2 Mississippi nonchilled 1000   31.5 0.03150000 10.00
#> 57   Mn3 Mississippi nonchilled   95   11.3 0.11894737  0.95
#> 58   Mn3 Mississippi nonchilled  175   19.4 0.11085714  1.75
#> 59   Mn3 Mississippi nonchilled  250   25.8 0.10320000  2.50
#> 60   Mn3 Mississippi nonchilled  350   27.9 0.07971429  3.50
#> 61   Mn3 Mississippi nonchilled  500   28.5 0.05700000  5.00
#> 62   Mn3 Mississippi nonchilled  675   28.1 0.04162963  6.75
#> 63   Mn3 Mississippi nonchilled 1000   27.8 0.02780000 10.00
#> 64   Mc1 Mississippi    chilled   95   10.5 0.11052632  0.95
#> 65   Mc1 Mississippi    chilled  175   14.9 0.08514286  1.75
#> 66   Mc1 Mississippi    chilled  250   18.1 0.07240000  2.50
#> 67   Mc1 Mississippi    chilled  350   18.9 0.05400000  3.50
#> 68   Mc1 Mississippi    chilled  500   19.5 0.03900000  5.00
#> 69   Mc1 Mississippi    chilled  675   22.2 0.03288889  6.75
#> 70   Mc1 Mississippi    chilled 1000   21.9 0.02190000 10.00
#> 71   Mc2 Mississippi    chilled   95    7.7 0.08105263  0.95
#> 72   Mc2 Mississippi    chilled  175   11.4 0.06514286  1.75
#> 73   Mc2 Mississippi    chilled  250   12.3 0.04920000  2.50
#> 74   Mc2 Mississippi    chilled  350   13.0 0.03714286  3.50
#> 75   Mc2 Mississippi    chilled  500   12.5 0.02500000  5.00
#> 76   Mc2 Mississippi    chilled  675   13.7 0.02029630  6.75
#> 77   Mc2 Mississippi    chilled 1000   14.4 0.01440000 10.00
#> 78   Mc3 Mississippi    chilled   95   10.6 0.11157895  0.95
#> 79   Mc3 Mississippi    chilled  175   18.0 0.10285714  1.75
#> 80   Mc3 Mississippi    chilled  250   17.9 0.07160000  2.50
#> 81   Mc3 Mississippi    chilled  350   17.9 0.05114286  3.50
#> 82   Mc3 Mississippi    chilled  500   17.9 0.03580000  5.00
#> 83   Mc3 Mississippi    chilled  675   18.9 0.02800000  6.75
#> 84   Mc3 Mississippi    chilled 1000   19.9 0.01990000 10.00

co2$transmutate(z2=~uptake/conc,y2=~conc/100)
co2
#>            z2    y2
#> 1  0.16842105  0.95
#> 2  0.17371429  1.75
#> 3  0.13920000  2.50
#> 4  0.10628571  3.50
#> 5  0.07060000  5.00
#> 6  0.05807407  6.75
#> 7  0.03970000 10.00
#> 8  0.14315789  0.95
#> 9  0.15600000  1.75
#> 10 0.14840000  2.50
#> 11 0.11942857  3.50
#> 12 0.08120000  5.00
#> 13 0.06133333  6.75
#> 14 0.04430000 10.00
#> 15 0.17052632  0.95
#> 16 0.18514286  1.75
#> 17 0.16120000  2.50
#> 18 0.12028571  3.50
#> 19 0.08580000  5.00
#> 20 0.06503704  6.75
#> 21 0.04550000 10.00
#> 22 0.14947368  0.95
#> 23 0.13771429  1.75
#> 24 0.12120000  2.50
#> 25 0.09885714  3.50
#> 26 0.06500000  5.00
#> 27 0.05244444  6.75
#> 28 0.03870000 10.00
#> 29 0.09789474  0.95
#> 30 0.15600000  1.75
#> 31 0.14000000  2.50
#> 32 0.11085714  3.50
#> 33 0.07720000  5.00
#> 34 0.05555556  6.75
#> 35 0.04240000 10.00
#> 36 0.15894737  0.95
#> 37 0.12000000  1.75
#> 38 0.15240000  2.50
#> 39 0.09714286  3.50
#> 40 0.07780000  5.00
#> 41 0.05866667  6.75
#> 42 0.04140000 10.00
#> 43 0.11157895  0.95
#> 44 0.10971429  1.75
#> 45 0.10480000  2.50
#> 46 0.08571429  3.50
#> 47 0.06180000  5.00
#> 48 0.04800000  6.75
#> 49 0.03550000 10.00
#> 50 0.12631579  0.95
#> 51 0.12571429  1.75
#> 52 0.12240000  2.50
#> 53 0.09085714  3.50
#> 54 0.06480000  5.00
#> 55 0.04607407  6.75
#> 56 0.03150000 10.00
#> 57 0.11894737  0.95
#> 58 0.11085714  1.75
#> 59 0.10320000  2.50
#> 60 0.07971429  3.50
#> 61 0.05700000  5.00
#> 62 0.04162963  6.75
#> 63 0.02780000 10.00
#> 64 0.11052632  0.95
#> 65 0.08514286  1.75
#> 66 0.07240000  2.50
#> 67 0.05400000  3.50
#> 68 0.03900000  5.00
#> 69 0.03288889  6.75
#> 70 0.02190000 10.00
#> 71 0.08105263  0.95
#> 72 0.06514286  1.75
#> 73 0.04920000  2.50
#> 74 0.03714286  3.50
#> 75 0.02500000  5.00
#> 76 0.02029630  6.75
#> 77 0.01440000 10.00
#> 78 0.11157895  0.95
#> 79 0.10285714  1.75
#> 80 0.07160000  2.50
#> 81 0.05114286  3.50
#> 82 0.03580000  5.00
#> 83 0.02800000  6.75
#> 84 0.01990000 10.00